Add Minables, Money, and Notifications
A notification text area with two lines was added to the center-top of the viewport. A Family of Minable stuff (meteors and satellites for now) were added which award Money upon collision. The main event sheet was organized with more folders for fuel, money and notifications. Invisible walls were added to a subfolder.