Improve Test Graphics and Notifications
Destroying/emptying of notification content was fixed by creating a separate function that checks the content of the display against the desired deletion. If they match, then the notification hasn't changed and it can be emptied. All of the current placeholder graphics were improved a bit to more accurately visualize the concept.