Add Day/Night Cycle and Altitude Display
The Player size was changed to 150x100. The Player's movement speeds were updated to match the new size. A text display that show the Player's current Altitude was added to the HUD. Another text display was added that shows the current in-game time as set by the new day/night cycle. That day/night cycle also adjusts the sky color.