Pivot from Construct3 to UE4
This project is no longer codenamed Rosebud and is now officially titled Cosmos Cleaner. The engine being used for development was switched from Construct 3 to Unreal Engine 4.27.2. This was possible by using a backup from Space Scavengers, an old game from university, and merging it with the nice menus of Minimal Match 3. This gave a head-start on the 2.5D game setup for Cosmos Cleaner.