Add Clutter
Created clutter that spawns in a random location, with random linear and angular velocity, and at random sizes. When the clutter moves off the edges of the play area, that clutter is destroyed and another clutter is spawned. Clutter does not collide with other clutter (this stops issues when spawning clutter). Also gravity is now turned off when above a certain altitude.