Improve Menu Flow
The settings menu was added to the in-game pause menu. That required a rewrite to the system that check whether the player is new or existing and whether the player was coming from in-game or is a menu-start user. Data should save between all menus. Also, the flow between menus was improved. When leaving in-game, the player now goes back to the main menu instead of the splash screen THEN the main menu.