Add Clutter Icons
Icons for the 12 clutter types were created in Adobe Illustrator and added to the HUD. Every clutter BP was created with proper attributes (except for size and speed). The game instance and savegame were renamed to include the game's name. The weight default value and HUD display issues were fixed.